requirements.txt
PyMySQL==1.0.2
holidays==0.11.1
xgboost~=1.3.3
requests==2.25.1
pandas~=1.2.5
Django==3.2.5
numpy~=1.20.2
scikit_learn==0.24.2
djangorestframework==3.12.4
django-debug-toolbar==3.2.1
setuptools~=52.0.0
scikit-learn~=0.24.2
django-environ==0.4.5
dateparser==1.0.0
pytz==2021.1